Add files via upload
Contains the Python Program that takes the various outputs of the MAPLE code and re formats them for later use by the SIngular package. Notice that these extra files will be generated only when we are calculating the Morse-theoretic flow-up basis of the assignment ring. For this to happen set the index Morse as 1 in the MAPLE code. See simplified-readme file for details.